refactor processing requirement
