+++ src/js/game/component_registry.js
@@ -15,7 +15,6 @@ import { ConstantSignalComponent } from "./components/constant_signal";
 import { LogicGateComponent } from "./components/logic_gate";
 import { LeverComponent } from "./components/lever";
 import { WireTunnelComponent } from "./components/wire_tunnel";
-import { ProcessingRequirementComponent } from "./components/processing_requirement";
 import { DisplayComponent } from "./components/display";
 
 export function initComponentRegistry() {
@@ -35,7 +34,6 @@ export function initComponentRegistry() {
     gComponentRegistry.register(LogicGateComponent);
     gComponentRegistry.register(LeverComponent);
     gComponentRegistry.register(WireTunnelComponent);
-    gComponentRegistry.register(ProcessingRequirementComponent);
     gComponentRegistry.register(DisplayComponent);
 
     // IMPORTANT ^^^^^ UPDATE ENTITY COMPONENT STORAGE AFTERWARDS
