+++ src/js/game/entity_components.js
@@ -15,7 +15,6 @@ import { ConstantSignalComponent } from "./components/constant_signal";
 import { LogicGateComponent } from "./components/logic_gate";
 import { LeverComponent } from "./components/lever";
 import { WireTunnelComponent } from "./components/wire_tunnel";
-import { ProcessingRequirementComponent } from "./components/processing_requirement";
 import { DisplayComponent } from "./components/display";
 /* typehints:end */
 
@@ -74,9 +73,6 @@ export class EntityComponentStorage {
 
         /** @type {WireTunnelComponent} */
         this.WireTunnel;
-
-        /** @type {ProcessingRequirementComponent} */
-        this.ProcessingRequirement;
 
         /** @type {DisplayComponent} */
         this.Display;
